pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "Currency exchange API"
authors = ["Kacper Akdag-Ochnik <kacperochnik01@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.3"
djangorestframework = "^3.15.2"
yfinance = "^0.2.50"
ipykernel = "^6.29.5"
drf-spectacular = "^0.27.2"
gunicorn = "^23.0.0"
whitenoise = "^6.8.2"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
